Fix player collision

+++ Source/Player/Player.cpp
@@ -160,14 +160,14 @@ void Player::update(float dt, World& world)
 
 void Player::collide(World& world, const glm::vec3& vel, float dt)
 {
-    for (float x = position.x - box.dimensions.x; x < position.x + box.dimensions.x; x++)
-    for (float y = position.y - box.dimensions.y; y < position.y + 0.7             ; y++)
-    for (float z = position.z - box.dimensions.z; z < position.z + box.dimensions.z; z++)
+    for (int x = position.x - box.dimensions.x; x < position.x + box.dimensions.x; x++)
+    for (int y = position.y - box.dimensions.y; y < position.y + 0.7             ; y++)
+    for (int z = position.z - box.dimensions.z; z < position.z + box.dimensions.z; z++)
     {
         auto block = world.getBlock(
-			static_cast<int>(x), 
-			static_cast<int>(y), 
-			static_cast<int>(z)
+			x, 
+			y, 
+			z
 		);
 
         if (block != 0 && block.getData().isCollidable)
